refactor: Remove v2 schema migration, fold changes into v1
