+++ src/handlers/browserTest/index.ts
@@ -20,7 +20,7 @@ export function initialize() {
     console.info(`${BrowserTestHandler.Create} event received`)
 
     const emptyBrowserTest: BrowserTestFile = {
-      version: '2.0',
+      version: '1.0',
       actions: [],
       settings: defaultBrowserTestOptions,
     }
