+++ src/test/factories/browserTest.ts
@@ -14,7 +14,7 @@ export function createBrowserTestFile(
   overrides?: Partial<BrowserTestFile>
 ): BrowserTestFile {
   return {
-    version: '2.0',
+    version: '1.0',
     actions: [],
     settings: defaultBrowserTestOptions,
     ...overrides,
